Add dynamic stl loader
