+++ packages/dev/loaders/src/dynamic.ts
@@ -1,9 +1,12 @@
 import { registerOBJLoader } from "./OBJ/objDynamicFileLoader";
 import { registerGLTFLoader } from "./glTF/glTFDynamicFileLoader";
 import { registerSPLATLoader } from "./SPLAT/splatDynamicFileLoader";
+import { registerSTLLoader } from "./STL/stlDynamicFileLoader";
 
+//test
 export function registerLoaders() {
     registerOBJLoader();
     registerGLTFLoader();
     registerSPLATLoader();
+    registerSTLLoader();
 }
